fixed the message and restructured the string utils
