+++ src/configuration/index.js
@@ -12,7 +12,7 @@ import renderTemplate from '../utils/render-template';
 import prepareReporters from '../utils/prepare-reporters';
 import WARNING_MESSAGES from '../notifications/warning-message';
 import log from '../cli/log';
-import getConcatenatedValuesString from '../utils/get-concatenated-values-string';
+import { getConcatenatedValuesString, getPluralSuffix } from '../utils/string';
 
 import {
     DEFAULT_TIMEOUT,
@@ -229,7 +229,7 @@ export default class Configuration {
             return;
 
         const optionsStr    = getConcatenatedValuesString(this._overridenOptions);
-        const optionsSuffix = this._overridenOptions.length > 1 ? 's' : '';
+        const optionsSuffix = getPluralSuffix(this._overridenOptions);
 
         Configuration._showConsoleWarning(renderTemplate(WARNING_MESSAGES.configOptionsWereOverriden, optionsStr, optionsSuffix));
