--- src/utils/get-concatenated-values-string.js
@@ -1,3 +1,0 @@
-export default function (array) {
-    return array.map(item => `"${item}"`).join(', ');
-}
